operator- of SciTrigger and SciEvent changed

+++ SCI_Decode/src/SciEvent.cpp
@@ -182,11 +182,11 @@ bool SciEvent::operator==(const SciEvent& right) const {
 }
 
 int SciEvent::operator-(const SciEvent& right) const {
-    return (period_ - right.period_) * CircleTime + (time_align - right.time_align);
+    return (period_ - right.period_) * CircleTime + (static_cast<int>(time_align) - static_cast<int>(right.time_align));
 }
 
 int SciEvent::operator-(const SciTrigger& right) const {
-    return (period_ - right.get_period()) * CircleTime + (time_align - right.time_align);
+    return (period_ - right.get_period()) * CircleTime + (static_cast<int>(time_align) - static_cast<int>(right.time_align));
 }
 
 void SciEvent::print(const Counter& cnt, ostream& os) {
